+++ tests/SimpleExcelReaderTest.php
@@ -150,4 +150,61 @@ class SimpleExcelReaderTest extends TestCase
             ],
         ], $rows);
     }
+
+    /** @test */
+    public function it_can_use_an_offset()
+    {
+        $rows = SimpleExcelReader::create($this->getStubPath('header-and-rows.csv'))
+            ->skip(1)
+            ->getRows()
+            ->toArray();
+
+        $this->assertEquals([
+            [
+                'email' => 'mary-jane@example.com',
+                'first_name' => 'mary jane',
+                'last_name' => 'doe',
+            ],
+        ], $rows);
+    }
+
+    /** @test */
+    public function it_can_take_a_limit()
+    {
+        $rows = SimpleExcelReader::create($this->getStubPath('header-and-rows.csv'))
+            ->take(1)
+            ->getRows()
+            ->toArray();
+
+        $this->assertEquals([
+            [
+                'email' => 'john@example.com',
+                'first_name' => 'john',
+                'last_name' => 'doe',
+            ],
+        ], $rows);
+    }
+
+    /** @test */
+    public function it_can_call_getRows_twice()
+    {
+        $reader = SimpleExcelReader::create($this->getStubPath('header-and-rows.csv'));
+        $firstRow = $reader->getRows()->first();
+        $firstRowAgain = $reader->getRows()->first();
+
+        $this->assertNotNull($firstRow);
+        $this->assertNotNull($firstRowAgain);
+    }
+
+    /** @test */
+    public function it_can_call_first_on_the_collection_twice()
+    {
+        $reader = SimpleExcelReader::create($this->getStubPath('header-and-rows.csv'));
+        $rowCollection = $reader->getRows();
+        $firstRow = $rowCollection->first();
+        $firstRowAgain = $rowCollection->first();
+
+        $this->assertNotNull($firstRow);
+        $this->assertNotNull($firstRowAgain);
+    }
 }