+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `simple-excel` will be documented in this file
 
+## 1.9.0 - 2020-10-30
+
+- add Header Styling Method (#39)
+
+## 1.8.1 - 2020-10-08
+
+- fix `skip` method
+
+## 1.8.0 - 2020-10-04
+
+- add 'take' and 'skip' functions to reader (#35)
+
 ## 1.7.1 - 2020-09-08
 
 - allow Laravel 8